backend/functions/requirements.txt
# Azure Functions
azure-functions>=1.12.0
azure-functions-durable>=1.2.0

# Azure Storage
azure-storage-blob>=12.14.0

# Azure AI Services
azure-ai-formrecognizer>=3.2.0
azure-core>=1.26.0

# Document Processing
PyPDF2>=3.0.0
python-docx>=0.8.11
pytesseract>=0.3.10
pdf2image>=1.16.3
tabula-py>=2.7.0
camelot-py>=0.10.1
opencv-python-headless>=4.7.0

# Data Processing
numpy>=1.24.0
pandas>=1.5.3

# HTTP and API
requests>=2.28.2
httpx>=0.23.3

# Utilities
python-dateutil>=2.8.2
pydantic>=1.10.7
